Project3/Project3/client/requirements.txt
pyodbc==5.1.0
rich==13.7.1
python-dateutil==2.9.0.post0 